pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "breakpoint_ai"
version = "2.0.0"
description = "Production-grade ATP Tennis Forecasting Pipeline"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "torch>=2.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",  # For strict config validation
    "tqdm>=4.66",
    "joblib>=1.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black",
    "isort",
    "jupyterlab"
]
